+++ src/checks.rs
@@ -1,3 +1,5 @@
 pub mod admin_role;
+pub mod check_user_exists;
+pub mod check_user_not_exists;
 pub mod enroll_channel;
 pub mod remove_role;
