+++ src/commands.rs
@@ -1,5 +1,6 @@
 pub mod acknowledgement;
 pub mod award;
+pub mod edit_user;
 pub mod enroll;
 pub mod help;
 pub mod ping;
