+++ src/storage.rs
@@ -1,3 +1,5 @@
+pub mod add_user_points;
 pub mod database_storage;
-pub mod edit_database;
+pub mod remove_user_points;
 pub mod user;
+pub mod user_exists;
